# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: FieldOps — Gestión de Órdenes de Trabajo
  version: 0.1.0
  description: >
    Contrato canónico (Principio II de la constitución — ruta fija /contracts/openapi.yaml).
    Cubre el slice: creación/(re)asignación de órdenes, registro de ejecución con evidencia,
    aprobación/rechazo, y listado por rol. Ver /specs/001-work-order-management/spec.md para
    los requisitos funcionales y no funcionales completos (FR-000..FR-020, NFR-01..NFR-08b).
    Incluye además login/refresh/logout (ver /specs/002-login-rbac/spec.md, FR-001..FR-010).
    Incluye además panel de dispatcher (listado/filtro/paginación de órdenes y técnicos,
    CRUD de clientes/técnicos, búsqueda de clientes, edición/cancelación de orden — ver
    /specs/003-dispatcher-orders-ui/spec.md, FR-001..FR-026).
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /technicians:
    get:
      operationId: listTechnicians
      summary: >
        Listar técnicos con estado activo/inactivo y cantidad de órdenes activas
        asignadas (003-dispatcher-orders-ui FR-011, FR-012) — rol dispatcher.
        Con `activo=true` sirve también al desplegable de asignación (FR-003, FR-004).
      parameters:
        - name: activo
          in: query
          required: false
          schema: { type: boolean }
          description: Filtra por estado activo/inactivo (usado por el desplegable de asignación con activo=true, FR-003/FR-004)
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, default: 50 }
          description: Tamaño fijo de página (003 FR-014) — el valor enviado por el cliente se ignora, siempre 50.
      responses:
        '200':
          description: Listado paginado de usuarios con rol tecnico
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedTechnicians' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      operationId: createTechnician
      summary: Crear un nuevo técnico (003-dispatcher-orders-ui FR-019) — rol dispatcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, email]
              properties:
                nombre: { type: string }
                email: { type: string, format: email }
      responses:
        '201':
          description: Técnico creado, activo por defecto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TechnicianSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: email ya registrado por otro usuario (FR-018, research.md §8)
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /technicians/{technicianId}:
    patch:
      operationId: updateTechnician
      summary: Editar nombre/email de un técnico existente (003-dispatcher-orders-ui FR-020) — rol dispatcher
      parameters:
        - name: technicianId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre: { type: string }
                email: { type: string, format: email }
      responses:
        '200':
          description: Técnico actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TechnicianSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: email ya registrado por otro usuario (FR-018, research.md §8)
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /technicians/{technicianId}/activo:
    parameters:
      - name: technicianId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: setTechnicianActive
      summary: Activar/desactivar técnico (Clarifications ronda 1 Q1; dispara FR-004d) — rol dispatcher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [activo]
              properties:
                activo: { type: boolean }
      responses:
        '200':
          description: >
            Estado actualizado. Si activo=false, toda orden en sin_asignar/pendiente_de_revision
            asignada a este técnico vuelve a sin_asignar (FR-004d), registrado en audit log.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: >
        JWT en cookie httpOnly (002-login-rbac Research §1 — nunca localStorage).
        Verificado por idpAdapter (001 Research §3, extendido). Ausente/inválido/expirado → 401.
        Fallo del validador externo (IdP/JWKS inalcanzable) → 503, nunca 401 (NFR-03b).
...
  responses:
    Unauthorized:
      description: Token ausente o inválido (NFR-03, FR-000)
    Forbidden:
      description: Token válido, rol sin permiso para la operación (FR-000, FR-019)
    NotFound:
      description: orderId inexistente, o rol con acceso potencial a la ruta sin acceso a esta orden (FR-019)
    ServiceUnavailable:
      description: >
        Fallo de dependencia externa crítica (KMS NFR-02b, IdP NFR-03b, audit log NFR-04b,
        antivirus NFR-08b) — la operación se aborta por completo, nunca degrada de forma insegura
  schemas:
...
    TechnicianSummary:
      description: Usuario rol tecnico + conteo derivado (003-dispatcher-orders-ui FR-011, FR-012)
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        activo: { type: boolean }
        activeOrderCount: { type: integer }
    PaginatedTechnicians:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TechnicianSummary' }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...